requirements.txt
contourpy==1.0.7
cycler==0.11.0
fonttools==4.39.4
importlib-resources==5.12.0
kiwisolver==1.4.4
matplotlib==3.7.1
numpy==1.24.3
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
pkg_resources==0.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-slugify==8.0.1
pytz==2023.3
seaborn==0.12.2
six==1.16.0
tabulate==0.9.0
text-unidecode==1.3
tzdata==2023.3
zipp==3.15.0
